Crate moonlander_ast

Models the moonlander problem using genetic programming.

The crate contains the AST node, fitness calculations, and reporting code (to interface with a web-based visualizer in another crate).

The code is exercised using the examples in the examples/ directory.

Modules

fitness
Calculate the fitness of moon landing programs (MODIFY HERE).

grammar
AST node types to express the programs that will land the moonlander (MODIFY HERE).

num
Numeric helpers.

run
Common evolution-running code between all examples.

sim
The moonlander simulation code.

Macros

println_err
Write text to stderr, so it doesn't interfere with the JSON output that gets written.