Function moonlander_ast::fitness::score_lander

pub fn score_lander<P>(program: &P, _: &mut Rng, sensor_data: SensorData, world: &World) -> LandingTrace where P: EvaluateToCommand + AstNode

Calculate the fitness of a single program.

This is done by running a simulation of the program until the moonlander hits the ground, and scoring the program over the course of the simulation.

Ideally, the fitness function should be as smooth as possible to guide the evolution in the direction of better solutions.

We return a full trace of all steps of the simulation so that we can visualize it using the program in the moonlander-visualization crate. The scores can be composed of multiple labeled sub-scores to gauge various parts of the score of a program.