Module moonlander_ast::grammar

AST node types to express the programs that will land the moonlander (MODIFY HERE).

Structs

CondList
Condition List node

Enums

Command
The command that the moonlander will execute.

CondChain
Condition list in the form of an AST node

Condition
A boolean condition, evaluating to either True or False.

Expression
Numeric expression.

Program
The root of a complex program that can decide between more than 2 cases.

Sensor
The various sensors that the moon lander program can use to base decisions on.