Module moonlander_ast::sim

The moonlander simulation code.

Modules

initial_state
Routines to generate starting positions for the moonlander.

Structs

LandingTrace
A trace of a moonlander's simulation and the corresponding score.

SensorData
SensorData represents the information that is available for programs to decide what Command to execute when it is evaluated, plus the values that are available to the fitness function to calculate a score for the program.

World
Various parameters to the simulation's physics.

Traits

BooleanValue
The truth value of a Condition.

EvaluateToCommand
Evaluate returns a specific Command to execute, depending on SensorData.

NumericValue
The numeric value of an Expression.

Functions

apply_command
Apply a command to the simulation.

evaluate_program
Evaluate a program and apply the resulting command to the simulation.